requirements.txt
rpi.gpio
lightdm
openbox
network-manager
python3-flask
python3-numpy
python3-requests
python3-gpiozero
python3-gunicorn
python3-pip
midori
unclutter
gunicorn3
